+++ src/Gara_Management/GUI/Item/itStockInDetail.xaml.cs
@@ -41,11 +41,11 @@ namespace Gara_Management.GUI.Item
             txtb_orderednum.Text=this.detail.GRNOrdinalNum.ToString();
             txtb_idStock.Text = this.detail.IDCom;
             check = true;
-            txtb_price.Text = this.detail.ComPrice.ToString();
+            txtb_price.Text =((int) this.detail.ComPrice).ToString();
             txtb_amount.Text=this.detail.ComQuantity.ToString();
             object nameCom = DataProvider.Instance.ExecuteScalar("SELECT NAME_COM from CAR_COMPONENTS WHERE ID_COM='" + this.detail.IDCom + "'");
             txtb_name.Text = nameCom.ToString();
-            txtb_sumofmoney.Text=this.detail.GRNTotalPayment.ToString();
+            txtb_sumofmoney.Text=((int)this.detail.GRNTotalPayment).ToString();
         }
 
         private void txtb_amount_TextChanged(object sender, TextChangedEventArgs e)
